% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliased_design.R
\docType{data}
\name{aliased_design}
\alias{aliased_design}
\title{D-efficient minimal aliasing design for five factors in 12 runs}
\format{
A tibble with 12 rows, 5 factors
}
\usage{
aliased_design
}
\description{
D-efficient minimal aliasing design for five factors in 12 runs
}
\keyword{datasets}
