% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plots.R
\name{contour_plots}
\alias{contour_plots}
\title{Contour plot(s) of a fitted linear model with ggplot2}
\usage{
contour_plots(rsm_contour, ncol = 2, return_list = FALSE)
}
\arguments{
\item{rsm_contour}{Contour using an rsm/lm object}

\item{ncol}{Specify the number of column used in grid layout for contour plots. Default is ncol=2}

\item{return_list}{If TRUE, a list of the generated contour plots will be returned. Otherwise, the generated contour plots will be displayed in a grid layout. Default is FALSE}
}
\value{
contours plot(s) of a fitted linear model with ggplot2. If return_list is TRUE, then a list of the individual contour plot(s) will be returned
}
\description{
Contour plot(s) of a fitted linear model with ggplot2
}
