% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_eff.R
\name{D_eff}
\alias{D_eff}
\title{D-efficiency for Composite Designs Based on Orthogonal Arrays}
\usage{
D_eff(
  X,
  design_type,
  subset = NA,
  ncenter = 0,
  efficiency = TRUE,
  lower_bounds = FALSE
)
}
\arguments{
\item{X}{Composite design matrix}

\item{design_type}{Specify between orthogonal array composite design (OACD) or central
composite design (CCD)}

\item{subset}{a subset of factors from our composite design}

\item{ncenter}{number of centers in our design}

\item{efficiency}{default is true,and returns D-efficiency. If false,
it will return D-optimal criteria instead}

\item{lower_bounds}{lower bounds of OACD D-efficiency}
}
\value{

}
\description{
D-efficiency for Composite Designs Based on Orthogonal Arrays
}
