% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_to_tibble.R
\name{design_to_tibble}
\alias{design_to_tibble}
\title{Convert an object of class `design` to `tibble`}
\usage{
design_to_tibble(design)
}
\arguments{
\item{design}{An object of class `design`}
}
\value{
Converted design to tibble
}
\description{
Convert an object of class `design` to `tibble`
}
\examples{
dat <- DoE.base::fac.design(factor.names = list(temp = c(16,32),
time = c(4,12)),replications = 5, randomize = FALSE)
Thk <- c(116.1, 106.7, 116.5, 123.2, 116.9, 107.5, 115.5, 125.1, 112.6, 105.9,
119.2, 124.5, 118.7, 107.1, 114.7, 124, 114.9, 106.5, 118.3, 124.7)
design <- DoE.base::add.response(dat, Thk)
design
design_to_tibble(design)
}
