% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Regression Diagnostic Plots with ggplot2}
\usage{
diagnostic_plots(
  model,
  standard_errors = FALSE,
  point_size = 1.5,
  theme_color = "#21908CFF",
  which_plots = 1:4,
  ncols = 2
)
}
\arguments{
\item{model}{Model of class "lm" or "glm"}

\item{standard_errors}{Display confidence interval around geom_smooth, FALSE by default}

\item{point_size}{Change size of points in plots}

\item{theme_color}{Change color of the geom_smooth line and text labels for the respective diagnostic plot}

\item{which_plots}{Choose which diagnostic plots to choose from. Options are 1 = 'residual vs fitted', 2 = 'Normal-QQ',
3 = 'Scale-location', 4 = 'Residual vs Leverage', 5 = "Cook's Distance". 6 = "Collinearity". Default is 1:4}

\item{ncols}{number of columns for grid layout. Default is 2}
}
\value{
Regression diagnostic plots
}
\description{
Regression Diagnostic Plots with ggplot2
}
\examples{
model <- lm(mpg ~ wt + am + gear + vs * cyl, data = mtcars)
diagnostic_plots(model,which_plots=1:6)
}
