% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitaxial.R
\docType{data}
\name{epitaxial}
\alias{epitaxial}
\title{Adapted epitaxial layer experiment}
\format{
A data.frame with 16 rows, 4 factors, and three responses to choose from
}
\source{
\url{http://www2.isye.gatech.edu/~jeffwu/book/data/adaptedlayer.dat}
}
\usage{
epitaxial
}
\description{
Adapted epitaxial layer experiment
}
\keyword{datasets}
