% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_boxplots.R
\name{gg_boxplots}
\alias{gg_boxplots}
\title{Boxplots using ggplot2}
\usage{
gg_boxplots(
  data,
  response,
  factor,
  horizontal = FALSE,
  point_size = 1,
  alpha = 1,
  color_palette = NA,
  direction = 1,
  show_mean = FALSE
)
}
\arguments{
\item{data}{provided dataset}

\item{response}{A character string indicating the response of the data}

\item{factor}{A character string indicating the factor of the data}

\item{horizontal}{Determine whether to change the orientation of the plot. Default is FALSE}

\item{point_size}{Change size of points (outliers) in boxplots}

\item{alpha}{The alpha transparency, a number in [0,1]}

\item{color_palette}{A character string indicating the color map option to use. Eight options are available: "viridis","cividis","magma","inferno","plasma","rocket","mako","turbo"}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest. If -1, the order of colors is reversed}

\item{show_mean}{Dispaly the mean for each boxplot. Default is FALSE}
}
\value{
Boxplots created with ggplot2
}
\description{
Boxplots using ggplot2
}
\examples{
gg_boxplots(data = throughput_dat,response = Throughput,factor = Machine)
gg_boxplots(data = throughput_dat,response = Throughput,factor = Machine,alpha = 0.7,
color_palette = 'viridis')
}
