% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_diagnostic_plots.R
\name{glm_diagnostic_plots}
\alias{glm_diagnostic_plots}
\title{GLM Diagnostic Plots with ggplot2}
\usage{
glm_diagnostic_plots(
  model,
  discrete_edm,
  which_plots = 1:4,
  ncols = 2,
  standard_errors = FALSE,
  theme_color = "#008EA0FF",
  point_size = 1.1,
  text_width = 25
)
}
\arguments{
\item{model}{Model of class "glm"}

\item{discrete_edm}{Logical value to exclusively specify if a discrete EDM is chosen to build model. Quantile residuals are used instead of Deviance/Pearson residuals for the plots for Discrete EDMs}

\item{which_plots}{Choose which diagnostic plots to choose from. \cr Options are 1 = "Residuals vs Fitted"; 2 = "Working Responses vs Linear Predictors"; 3 = "Normal Q-Q"; 4 = "Outlier Detection"; 5 = "Half norm plot using leverages"; 6 = "Half norm plot using Cook's Distance"; 7 = "Cook's Distance"; 8 = "DFFITS"; 9 = "VIF"}

\item{ncols}{number of columns for grid layout. Default is 2}

\item{standard_errors}{Display confidence interval around geom_smooth, FALSE by default}

\item{theme_color}{Change color of the geom_smooth line and text labels for the respective diagnostic plot}

\item{point_size}{Change size of points in plots}

\item{text_width}{Passed onto `stringr::str_wrap`. Positive integer giving target line width in characters. A width less than or equal to 1 will put each word on its own line.}
}
\value{
GLM diagnostic plots
}
\description{
GLM Diagnostic Plots with ggplot2
}
\details{
Plots 1 & 2: Check the assumptions for the *systematic component of the GLM*.

Plot 3: Check the for the *random component (ie, EDM distribution) of the GLM*.

Plot 4: Check for *outliers* with |residuals| > 3.

Plots 5 & 6: Check for outliers with *half norm plots*.

Plot 7: Check for outliers with *Cook's Distance*. A data point having a large Cook's distance indicates that the data point strongly influences the fitted values of the model. The threshold used for detecting or classifying observations as outers is 4/n4/n where nn is the number of observations.

Plot 8: Check for outliers with *DFFITS*, which considers how much the fitted value of observation i changes between the model fitted with all the data and the model fitted with obervation i omitted. The threshold used for detecting or classifying observations as outers is \eqn{2 * sqrt((parameters + 1) / (observations - parameters - 1))}

Plot 9: Check for collinearity with the variance inflation factor (VIF). Tolerance = \eqn{1 - R_j^2}, VIF = (1/Tolerance) where R_j^2 is the coefficient of determination of a regression of predictor jj on all the other predictors. A general rule of thumb is that VIFs exceeding 4 warrant further investigation, while VIFs exceeding 10 indicates a multicollinearity problem
}
\examples{
model <- glm(Volume ~ Girth + Height, family = Gamma(link = "log"), data = trees)

# Default plots returned
glm_diagnostic_plots(model, discrete_edm = FALSE, which_plots = 1:4)

# Half norm plots
glm_diagnostic_plots(model, discrete_edm = FALSE, which_plots = 5:6)

# Cook's Distance & DFFITS
glm_diagnostic_plots(model, discrete_edm = FALSE, which_plots = 7:8)

# Collinearity
glm_diagnostic_plots(model, discrete_edm = FALSE, which_plots = 9, ncols = 1)
}
