% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IA_effect.R
\name{IA_effect}
\alias{IA_effect}
\title{Obtain interaction effects plot between two factors in our design with ggplot}
\usage{
IA_effect(data, factr1, factr2, response)
}
\arguments{
\item{response}{}
}
\value{
data to calculate interaction effects, and interaction effects plot
}
\description{
Obtain interaction effects plot between two factors in our design with ggplot
}
\examples{
IA_effect(data = epitaxial, factr1 = A, factr2 = B, response = ybar)
}
