% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_effect.R
\name{interaction_effect}
\alias{interaction_effect}
\title{Obtain interaction effects plot between two factors in a factorial design}
\usage{
interaction_effect(
  data,
  factor_1,
  factor_2,
  response_var,
  linetypes = c("solid", "dashed"),
  colors = c("#4260c9", "#d6443c"),
  ...
)
}
\arguments{
\item{data}{Data/Design used to plot}

\item{factor_1}{First factor}

\item{factor_2}{Second factor}

\item{response_var}{Response variable}

\item{linetypes}{Change linetypes. Default are ('solid','dashed)}

\item{colors}{Change color of lines/points. Default are ("#4260c9" ,"#d6443c")}

\item{...}{additional customization for both geom_line and geom_point (optional)}
}
\value{
interaction effects plot
}
\description{
Obtain interaction effects plot between two factors in a factorial design
}
\examples{
interaction_effect(data = adapted_epitaxial, factor_1 = A,
factor_2 = B, response_var = ybar)
}
