% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_effect.R
\name{main_effect}
\alias{main_effect}
\title{Obtain main effect plot for a factor in a factorial design}
\usage{
main_effect(
  data,
  factor_1,
  response_var,
  colour = "#4260c9",
  point_size = 1.5,
  line_size = 0.8,
  linetype = "solid",
  shape = 16
)
}
\arguments{
\item{data}{Data/Design used to plot}

\item{factor_1}{Factor variable (without quotes)}

\item{response_var}{Response variable (without quotes)}

\item{colour}{A character string specifying the color of line/points}

\item{point_size}{Change size of points}

\item{line_size}{Change thickness of line}

\item{linetype}{Change type of line. Default is 'solid'}

\item{shape}{Change shape of points. Default is 16}
}
\value{
main effects plot
}
\description{
Obtain main effect plot for a factor in a factorial design
}
\examples{
main_effect(data = adapted_epitaxial, factor_1 = B, response_var = ybar)
}
