% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_epitaxial.R
\docType{data}
\name{original_epitaxial}
\alias{original_epitaxial}
\title{Original epitaxial layer experiment}
\format{
A tibble with 16 rows, 4 factors (A,B,C,D), and three responses to choose from (ybar,s2,lns2)
}
\usage{
original_epitaxial
}
\description{
Original epitaxial layer experiment
}
\keyword{datasets}
