% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pulp_experiment.R
\docType{data}
\name{pulp_experiment}
\alias{pulp_experiment}
\title{Reflectance Data, Pulp Experiment}
\format{
A tibble with 5 rows, and 4 variables (A,B,C,D)
}
\source{
Wu, CF Jeff, and Michael S. Hamada. Experiments: planning, analysis, and optimization. John Wiley & Sons, 2011
}
\usage{
pulp_experiment
}
\description{
Plant performance is based on pulp brightness as measured by a reflectance meter.
Each of the four shift operators (denoted by A, B, C, and D) made five pulp handsheets from unbleached pulp.
Reflectance was read for each of the handsheets using a brightness tester
}
\keyword{datasets}
