% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{second_order_mat}
\alias{second_order_mat}
\title{Second Order Model Design Matrix}
\usage{
second_order_mat(model, terms = "full")
}
\arguments{
\item{model}{Input a second order model}

\item{terms}{specify if you only want to extract certain terms from the second order design matrix}
}
\value{
A design matrix from initial second order model
}
\description{
Obtain second order design matrix from a specified model. Optionally, be
able to extract specific terms from the matrix (i.e linear terms, quadratic terms, or bilinear terms)
}
\examples{
model2 <- lm(ybar ~ (A+B)^2 + I(A^2)+I(B^2), data= AEL_dat)
second_order_mat(model2)
second_order_mat(model2, terms = "linear")
second_order_mat(model2, terms = "quadratic")

}
