% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summs_models.R
\name{summs_models}
\alias{summs_models}
\title{Compare factor effects from multiple models and visualize
the difference in absolute value T statistics to check for model performances}
\usage{
summs_models(models, model.names, showplot = TRUE)
}
\arguments{
\item{models}{a list of models to compare}

\item{model.names}{a vector of names for the models in format of "model_name"}

\item{showplot}{logical indicating to show the plot. If false, only the data used to construct the plot is returned. Default is TRUE}
}
\value{
summary output between models, and ggplot to compare models' consistency
}
\description{
Compare factor effects from multiple models and visualize
the difference in absolute value T statistics to check for model performances
}
\examples{
m1 <- lm(ybar~(A+B+C)^2,data =adapted_epitaxial)
m2 <- lm(ybar~(A+B)^2,data =adapted_epitaxial)

models <- list(m1,m2)
models.names <- c("Model1","Model2")
summs_models(models,models.names)
}
