% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bw_nogrid.R
\name{theme_bw_nogrid}
\alias{theme_bw_nogrid}
\title{Theme for plots used in 'ggDoE'}
\usage{
theme_bw_nogrid()
}
\value{
A simple black and white theme without grid.major and grid.minor for ggplot objects.
}
\description{
Theme for plots used in 'ggDoE'
}
\examples{
library(ggplot2)
data <- ToothGrowth
data$dose <- factor(data$dose,levels = c(0.5, 1, 2),
                    labels = c("D0.5", "D1", "D2"))

ggplot(data, aes(x=dose, y=len)) +
 geom_boxplot()+
 theme_bw_nogrid()
}
