% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throughput_dat.R
\docType{data}
\name{throughput_dat}
\alias{throughput_dat}
\title{Throughput_dat}
\format{
A tibble with 25 rows, 4 factors (Day,Operator,Machine,Method), and one response (Throughput)
}
\source{
\url{https://www2.isye.gatech.edu/~jeffwu/wuhamadabook/data/throughput.dat}
}
\usage{
throughput_dat
}
\description{
Throughput_dat
}
\keyword{datasets}
