% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoD_projections.R
\name{twoD_projections}
\alias{twoD_projections}
\title{Two Dimensional Projections of Latin Hypercube Designs}
\usage{
twoD_projections(
  design,
  point_color = "#21908CFF",
  grid = FALSE,
  point_size = 1.5,
  ncols = 2
)
}
\arguments{
\item{design}{A latin hypercube design. Can be matrix, or data.frame}

\item{point_color}{Change color of points in plots}

\item{grid}{Logial argument. Specify if a grid should be added to each projection.
The grid is constructed using seq(0,1,length=nrows(design)+1). Default is FALSE}

\item{point_size}{Change size of points in plots}

\item{ncols}{number of columns for grid layout. Default is 2}
}
\value{
A grid of ggplots from all two dimensional projections of a latin hypercube design.
}
\description{
Two Dimensional Projections of Latin Hypercube Designs
}
\examples{
set.seed(100)
X <- lhs::randomLHS(n = 10, k = 4)
twoD_projections(X,ncols=3,grid = TRUE)
}
