% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapted_epitaxial.R
\docType{data}
\name{adapted_epitaxial}
\alias{adapted_epitaxial}
\title{Adapted epitaxial layer experiment}
\format{
A tibble with 16 rows, 4 factors (A,B,C,D), and three responses to choose from (ybar,s2,lns2)
}
\source{
\url{http://www2.isye.gatech.edu/~jeffwu/book/data/adaptedlayer.dat}
}
\usage{
adapted_epitaxial
}
\description{
Adapted epitaxial layer experiment
}
\keyword{datasets}
