% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-adapted_epitaxial.R
\docType{data}
\name{adapted_epitaxial}
\alias{adapted_epitaxial}
\title{Adapted epitaxial layer experiment}
\format{
A tibble with 16 rows, 4 factors (A,B,C,D), and three responses to choose from (ybar,s2,lns2)
}
\usage{
adapted_epitaxial
}
\description{
Adapted epitaxial layer experiment
}
\keyword{datasets}
