% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Regression Diagnostics in ggplot2}
\usage{
diagnostic_plots(model, se = TRUE, point_size = 3.5)
}
\arguments{
\item{model}{Model of class "lm" or "glm"}

\item{se}{Display confidence interval around smooth? TRUE by default}

\item{point_size}{Change size of points in plots}
}
\value{
Regression Diagnostic plots
}
\description{
Regression Diagnostics in ggplot2
}
\examples{
epitaxial_ybar <- lm(ybar ~ A+B+C+D,data=epitaxial)
diagnostic_plots(epitaxial_ybar)
}
