% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plots.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Regression Diagnostic Plots with ggplot2}
\usage{
diagnostic_plots(model, SE = FALSE, point_size = 3, line_color = "indianred3")
}
\arguments{
\item{model}{Model of class "lm" or "glm"}

\item{SE}{Display confidence interval around geom_smooth, FALSE by default}

\item{point_size}{Change size of points in plots}

\item{line_color}{Change color of the geom_smooth line for the respective diagnostic plot}
}
\value{
Regression Diagnostic plots. In the case where all hat values are equal only residual vs.fitted, normal-QQ, and scale-location plots are returned
}
\description{
Regression Diagnostic Plots with ggplot2
}
\examples{
epitaxial_ybar <- lm(ybar ~ A+B+C+D,data=original_epitaxial)
diagnostic_plots(epitaxial_ybar)
}
