% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_normal.R
\name{half_normal}
\alias{half_normal}
\title{Half-Normal Effects Plots}
\usage{
half_normal(
  obj,
  method = "Lenth",
  alpha = 0.05,
  label_active = FALSE,
  ref_line = FALSE,
  margin_errors = FALSE,
  showplot = TRUE
)
}
\arguments{
\item{obj}{object of class lm}

\item{method}{Character value. Method to calculate PSE. Default is Lenth. Options include: Zahn, WZahn, Lenth, RMS, Dong, JuanPena, Daniel}

\item{alpha}{specify the significance level. Default is alpha=0.05}

\item{label_active}{If TRUE, active effects are labeled if the effects cross the computed margin of error (ME). See method argument for more details}

\item{ref_line}{Dafault is TRUE, if FALSE the abline with slope (1/PSE) is not displayed. Reference line should follow along most points that are not considered outliers.}

\item{margin_errors}{Default is FALSE, if TRUE the cutoffs for margin of errors (ME) and simultaneous margin of error (SME) are shown}

\item{showplot}{Default is TRUE, if FALSE plot will not be shown and a tibble is returned used to create the plot along with the calculated PSE,ME,SME}
}
\value{
A tibble with the absolute effects and half-normal quantiles. A ggplot2 version of halfnormal plot for factorial effects is returned
}
\description{
Half-Normal Effects Plots
}
\examples{
m1 <- lm(lns2 ~ (A+B+C+D)^4,data=original_epitaxial)
half_normal(m1)
half_normal(m1,alpha=0.1,label_active=TRUE,margin_errors=TRUE)
half_normal(m1,method='Zahn',alpha=0.1,ref_line=TRUE,label_active=TRUE,margin_errors=TRUE)
}
