% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfnormal.R
\name{halfnormal}
\alias{halfnormal}
\title{Half-Normal Effects Plots}
\usage{
halfnormal(obj, alpha = 0.05, signif_label = FALSE)
}
\arguments{
\item{obj}{object of class lm}

\item{alpha}{significance level for the Lenth method. Default is 0.05}

\item{signif_label}{If TRUE, only the significant factors according to the Lenth method
(significance level given by alpha) are labeled}
}
\value{
A dataframe with the absolute effects and half-normal qunatiles.
A ggplot version of halfnormal plot for factorial effects is returned
}
\description{
Half-Normal Effects Plots
}
\examples{
m1 <- lm(lns2 ~ (A+B+C+D)^4,data=epitaxial)
halfnormal(m1)
halfnormal(m1,alpha=0.1)
halfnormal(m1,alpha=0.2,signif_label=TRUE)
}
