% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_effect.R
\name{interaction_effect}
\alias{interaction_effect}
\title{Obtain interaction effects plot between two factors in a factorial design}
\usage{
interaction_effect(
  Design,
  factr1,
  factr2,
  response,
  point_size = 1.5,
  line_size = 1,
  showplot = TRUE
)
}
\arguments{
\item{Design}{Data/Design used to plot}

\item{factr1}{First factor}

\item{factr2}{Second factor}

\item{response}{Response variable}

\item{point_size}{Change size of points}

\item{line_size}{Change thickness of lines}

\item{showplot}{logical indicating to show main effect plot. If false, a tibble containing the data used to construct the plot is returned. Default is TRUE}
}
\value{
data to calculate interaction effects, and interaction effects plot
}
\description{
Obtain interaction effects plot between two factors in a factorial design
}
\examples{
interaction_effect(Design = adapted_epitaxial, factr1 = A, factr2 = B, response = ybar)
}
