% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lenth_method.R
\name{Lenth_method}
\alias{Lenth_method}
\title{Lenth’s method: testing effect significance for experiments without variance estimates}
\usage{
Lenth_method(mod, alpha = 0.05)
}
\arguments{
\item{mod}{A linear model}

\item{alpha}{specify the significance level. Default is alpha=0.05}
}
\value{
PSE,ME,SME using Lenth's method. Also returns a plot visualizing any factor which is significant using the obtained
values of ME, and SME.
}
\description{
Lenth’s method: testing effect significance for experiments without variance estimates
}
\examples{
m1 <- lm(ybar ~ (A+B+C+D)^2,data=epitaxial)
Lenth_method(m1)
Lenth_method(m1,alpha=0.01)
}
