% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_effect.R
\name{main_effect}
\alias{main_effect}
\title{Obtain main effect plot for a factor in our design with ggplot}
\usage{
main_effect(data, factr, response)
}
\arguments{
\item{data}{Data/Design used to plot}

\item{factr}{Factor}

\item{response}{Response variable}
}
\value{
data to calculate main effect, and main effect plot
}
\description{
Obtain main effect plot for a factor in our design with ggplot
}
\examples{
main_effect(data=epitaxial, factr = B, response = ybar)
}
